Implements Prediction Mask for Pixel Classification
This still doesn;t work as a backwards compatible macro,
since using pixel classification as a macro will require the
value for 'useMask'.

I could not find a way to have this extra parameter with a default
value that would be picked up also by the macro usage. Either
the input is marked as resolved and doesn't show up in the UI,
or it is not marked as resolved, but requires the macro usage to specify
the value interactively.